moved stores to lib/data folder
